"check_style_score -> 5 of 5 passed. style score: 0/1 "

+++ cspp2-practice/M2/Assignment-3/power.java
@@ -7,11 +7,11 @@ public class power {
 /* Fill the main function to print the number of 7's between 1 to n*/
     public static void main(String[] args) {
         Scanner scan = new Scanner(System.in);
-        int x = scan.nextInt();
-        int y = scan.nextInt();
+        long x = scan.nextInt();
+        long y = scan.nextInt();
         System.out.println(power(x, y));
     }
-        static int power(int x, int y){
+        static long power(long x, long y){
             if (y == 0)
                 return 1;
             else if (y == 1)
